Successful Overhaul
Phew! *wipes sweat off forehead*

Okay, so;
- GOT IT RUNNING!
- It's loading chunks from data file properly, figured out that we had
to convert it from using your player vector (pos(x,y,z)) to the chunk
vector (pos = playerToChunk(pos))
- Dead slow framerate, but it is a constant framerate which is better
than what it was before where it was one frame every minute roughly lol
- Found the new problem:

*************************************
********THE NEW PROBLEM********
*************************************
We need to rework all of the code. Sebastian pointed out that in
Minecraft, they use a chunk that is 16x*16z*256z. SO. It's not a cube,
it's actually a rectangle. This poses many problems, yes, but I think it
will actually be pretty simple to rework out code to do that.
In short:
- We need to rework code to eliminate the (Y) coordinate from the chunk
system. The chunk system will only need (X,Z) for locations! This will
drastically increase performance and also logic (making it much easier
to create our chunks themselves so it's not all almost the same chunk!
:D)
- Extend the generation of (Y) blocks to be from 0 to 255. (inclusive
that would be 256 blocks high (16*16, so it's still a factor of 2(better
performance))), but leave X and Y as being 16.

I'll continue working on this tonight and throw my attempt at this
unless you text me/email me that you've synced anything new or are doing
it yourself!